Issue 20, Hardware volume buttons should work from all activities
