+++ src/com/danga/squeezer/SqueezerRandomplayActivity.java
@@ -19,28 +19,20 @@ package com.danga.squeezer;
 
 import java.util.Arrays;
 
-import android.content.ComponentName;
 import android.content.Context;
 import android.content.Intent;
-import android.content.ServiceConnection;
 import android.os.Bundle;
-import android.os.IBinder;
 import android.os.RemoteException;
-import android.support.v4.app.FragmentActivity;
 import android.util.Log;
 import android.view.View;
 import android.widget.AdapterView;
-import android.widget.ListView;
 import android.widget.AdapterView.OnItemClickListener;
+import android.widget.ListView;
 
+import com.danga.squeezer.framework.SqueezerBaseActivity;
 import com.danga.squeezer.menu.SqueezerMenuFragment;
-import com.danga.squeezer.service.ISqueezeService;
-import com.danga.squeezer.service.SqueezeService;
 
-public class SqueezerRandomplayActivity extends FragmentActivity {
-	private static final String TAG = SqueezerRandomplayActivity.class.getName();
-
-	private ISqueezeService service;
+public class SqueezerRandomplayActivity extends SqueezerBaseActivity {
     private ListView listView;
 
 	@Override
@@ -53,28 +45,9 @@ public class SqueezerRandomplayActivity extends FragmentActivity {
 	}
 
     @Override
-    public void onResume() {
-        super.onResume();
-        bindService(new Intent(this, SqueezeService.class), serviceConnection, Context.BIND_AUTO_CREATE);
-        Log.d(TAG, "did bindService; serviceStub = " + service);
+    protected void onServiceConnected() throws RemoteException {
     }
 
-	private final ServiceConnection serviceConnection = new ServiceConnection() {
-		public void onServiceConnected(ComponentName name, IBinder binder) {
-            service = ISqueezeService.Stub.asInterface(binder);
-        }
-        public void onServiceDisconnected(ComponentName name) {
-            service = null;
-        };
-    };
-
-	@Override
-    public void onPause() {
-        super.onPause();
-        if (serviceConnection != null) {
-        	unbindService(serviceConnection);
-        }
-    }
 
 	private void setRandomplayMenu() {
 		String[] values = getResources().getStringArray(R.array.randomplay_items);
@@ -89,9 +62,9 @@ public class SqueezerRandomplayActivity extends FragmentActivity {
 		public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
 			if (position < RandomPlayType.values().length) {
 				try {
-					service.randomPlay(RandomPlayType.values()[position].toString());
+					getService().randomPlay(RandomPlayType.values()[position].toString());
 				} catch (RemoteException e) {
-	                Log.e(TAG, "Error registering list callback: " + e);
+	                Log.e(getTag(), "Error registering list callback: " + e);
 				}
 				SqueezerActivity.show(SqueezerRandomplayActivity.this);
 				return;
