+++ src/com/danga/squeezer/framework/SqueezerBaseActivity.java
@@ -25,6 +25,7 @@ import android.os.IBinder;
 import android.os.RemoteException;
 import android.support.v4.app.FragmentActivity;
 import android.util.Log;
+import android.view.KeyEvent;
 
 import com.danga.squeezer.service.ISqueezeService;
 import com.danga.squeezer.service.SqueezeService;
@@ -86,6 +87,54 @@ public abstract class SqueezerBaseActivity extends FragmentActivity {
 	public Handler getUIThreadHandler() {
 		return uiThreadHandler;
 	}
+
+	
+    /*
+     * Intercept hardware volume control keys to control Squeezeserver
+     * volume.
+     *
+     * Change the volume when the key is depressed.  Suppress the keyUp
+     * event, otherwise you get a notification beep as well as the volume
+     * changing.
+     */
+    @Override
+    public boolean onKeyDown(int keyCode, KeyEvent event) {
+        switch (keyCode) {
+        case KeyEvent.KEYCODE_VOLUME_UP:
+            changeVolumeBy(+5);
+            return true;
+        case KeyEvent.KEYCODE_VOLUME_DOWN:
+            changeVolumeBy(-5);
+            return true;
+        }
+
+        return super.onKeyDown(keyCode, event);
+    }
+
+    @Override
+    public boolean onKeyUp(int keyCode, KeyEvent event) {
+        switch (keyCode) {
+        case KeyEvent.KEYCODE_VOLUME_UP:
+        case KeyEvent.KEYCODE_VOLUME_DOWN:
+            return true;
+        }
+
+        return super.onKeyUp(keyCode, event);
+    }
+
+    private boolean changeVolumeBy(int delta) {
+        if (getService() == null) {
+            return false;
+        }
+        Log.v(getTag(), "Adjust volume by: " + delta);
+        try {
+            getService().adjustVolumeBy(delta);
+            return true;
+        } catch (RemoteException e) {
+            Log.e(getTag(), "Error from service.adjustVolumeBy: " + e);
+        }
+        return false;
+    }
 
 
 	// This section is just an easier way to call squeeze service
